refactored Redux Example (#819)
