+++ examples/ReduxExample/src/reducers/index.js
@@ -1,0 +1,43 @@
+import { combineReducers } from 'redux';
+import { NavigationActions } from 'react-navigation';
+
+import { AppNavigator } from '../navigators/AppNavigator';
+
+const initialNavState = {
+  index: 1,
+  routes: [
+    { key: 'InitA', routeName: 'Main' },
+    { key: 'InitB', routeName: 'Login' },
+  ],
+};
+
+const initialAuthState = { isLoggedIn: false };
+
+function nav(state = initialNavState, action) {
+  switch (action.type) {
+    case 'Login':
+      return AppNavigator.router.getStateForAction(NavigationActions.back(), state);
+    case 'Logout':
+      return AppNavigator.router.getStateForAction(NavigationActions.navigate({ routeName: 'Login' }), state);
+    default:
+      return AppNavigator.router.getStateForAction(action, state);
+  }
+}
+
+function auth(state = initialAuthState, action) {
+  switch (action.type) {
+    case 'Login':
+      return { ...state, isLoggedIn: true };
+    case 'Logout':
+      return { ...state, isLoggedIn: false };
+    default:
+      return state;
+  }
+}
+
+const AppReducer = combineReducers({
+  nav,
+  auth,
+});
+
+export default AppReducer;
